TABP-28 Refactor entities and repository for improved functionality

+++ TABP/TABP.Domain/Entities/Amenity.cs
@@ -1,9 +1,9 @@
 ﻿using TABP.Domain.Entities.Common;
 namespace TABP.Domain.Entities
 {
-   public class Amenity: SoftDeletable
+   public class Amenity : SoftDeletable
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }= null!;
         public string Description { get; set; } = null!;
         public DateTime CreatedAt { get; set; }
